Вынес скрапинг в отдельный тред
